# app/static/swagger.yaml
...
swagger: '2.0'
info:
  title: "Mechanic API"
  description: "A mechanic shop API recording customers, mechanics, 
                service tickets, inventory, and their relationships"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apikey
    name: Authorization
    in: header


# Endpoints
paths:
...
  /mechanics:
    post:
      tags:
        - Mechanics
      summary: "Endpoint to create a new Mechanic"
      description: "Endpoint that creates a new mechanic,
                    email must be unique"
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new mechanic"
          required: true
          schema:
            $ref: "#/definitions/CreateMechanicPayload"
      responses:
        201:
          description: "Successful Mechanic Creation"
          schema:
            $ref: "#/definitions/CreateMechanicResponse"
          examples:
            application/json:
              id: 1
              name: Bart McConnan
              email: mconnan@email.com
              phone: 123-456-7890
              salary: 45000.0
              status: "success"
        400:
          description: "Email already associated with an account"
          schema:
            $ref: "#/definitions/CreateMechanicResponse"
          examples:
            application/json:
              message: "Email already asociated with an account"
              status: "bad request"

  # Get All Mechanics
    get:
      tags:
        - Mechanics
      summary: "Endpoint that returns all mechanics"
      descruption: "Endpoint that retieves a list of all mechanics"
      responses:
        200:
          description: "Retrieved Mechanics Successfully"
          schema:
            $ref: "#/definitions/AllMechanics"
            status: "success"

  # Update a Mechanic
    put:
      tags:
        - Mechanics
      summary: "Endpoint to update a Mechanic"
      description: "Endpoint that updates a mechanic's information."
      parameters:
        - in: "body"
          name: "body"
          description: "Information needed to update a Mechanic"
          required: true
          schema:
            $ref: "#/definitions/UpdateMechanicPayload"
      responses:
        200:
          description: "Updated Mechanic Successfully"
          schema:
            $ref: "#/definitions/UpdateMechanicResponse"
          examples:
            application/json:
              id: 1
              name: Bart McConnan
              email: mconnan@email.com
              phone: 123-456-7890
              salary: 45000.0
              status: "success"  
        404:
          description: "Mechanic not found"             
          schema:
            $ref: "#/definitions/UpdateMechanicResponse"
          examples:
            application/json: 
              message: "Mechanic not found"
              status: "bad request"
        400: 
          description: "Validation Error"
          schema:
            $ref: "#/definitions/UpdateMechanicResponse"
          examples:
            application/json: 
              message: "<Validation Error>"
              status: "bad request"

  # Delete a Mechanic
    delete:
      tags:
        - Mechanics
      summary: "Endpoint to delete a Mechanic"
      description: "Endpoint that deletes a Mechanic using known mechanic id"
      responses:
        200:
          description: "Deleted Mechanic Successfully"
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
          examples:
            application/json:
              message: "Mechanic <mechanic_id>, successfully deleted"
              status: "success"
        404:
          description: "Mechanic not found"
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
          examples:
            application/json:
              message: "Mechanic not found"
              status: "bad request"

  # Get Specific Mechanic
  /mechanics/<int:mechanic_id>:
    get:
      tags:
        - Mechanics
      summary: "Endpoint to return a specific mechanic"
      description: "Endpoint that retrieves a mechanic using the known mechanic_id"
      responses:
        200:
          description: "Retrieved Mechanic Successfully"
          schema:
            $ref: "$/definitions/SpecificMechanic"
          examples:
            application/json:
              id: 1
              name: Bart McConnan
              email: mconnan@email.com
              phone: 123-456-7890
              salary: 45000.0
              status: "success" 
        404:
          description: "Mechanic not found"
          schema:
            $ref: "$/definitions/SpecificMechanic"
          examples:
            application/json:
              message: "Mechanic not found"
              status: "bad request"               

  # Get List of Mechanics in order of who has worked on the most service tickets
...
definitions: 
...
  DeleteCustomerResponse:
    type: "object"
    properties:
      message:
        type: "string"

  # GET Service Tickets for a Customer (includes mechanics working on ticket)
...
  CreateMechanicPayload:
    type: "object"
    properties: 
      name: 
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary: 
        type: "float"

    required:
      - name
      - email
      - phone
      - salary
  CreateMechanicResponse:
    type: "object"
    properties: 
      id:
        type: "integer"
      name: 
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary: 
        type: "float"

  # Get all Mechanics
  AllMechanics:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name: 
          type: "string"
        email:
          type: "string"
        phone:
          type: "string"
        salary:
          type: "float"

  # Get a Specific Mechanic
...
  UpdateMechanicPayload:
    type: "object"
    properties: 
      name: 
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary: 
        type: "float"

      required:
        - name
        - email
        - phone
        - salary
  UpdateMechanicResponse:
    type: "object"
    properties: 
      id:
        type: "integer"
      name: 
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary: 
        type: "float"

  # Delete a Mechanic
...